+++ src/components/User/Assets/AssetRequests/AssetRequests.Show.vue
@@ -161,6 +161,10 @@
               assetRequest,
               'operationDetails.creatorDetails.logo.key'
             )"
+            :mime-type="safeGet(
+              assetRequest,
+              'operationDetails.details.logo.mime_type'
+            )"
           />
         </span>
       </div>